TF-TRT - Move additional utils to trt_testutils.
